tests/wpt/tests/css/css-transforms/2d-rotate-js.html

<!DOCTYPE html>
<html>
	<head>
		<title>JS test: Rotate via javascript must show the correct computed rotation</title>
		<link rel="author" title="Rick Hurst" href="http://mrkn.co/axegs">
		<link rel="help" href="https://drafts.csswg.org/css-transforms-1/#serialization-of-the-computed-value">
		<link rel="help" href="https://drafts.csswg.org/css-transforms-1/#serialization-of-transform-functions">
		<script src="/resources/testharness.js"></script>
		<script src="/resources/testharnessreport.js"></script>
		<meta name="flags" content="svg">
		<meta name="assert" content="Asserting that you can rotate an element with JS and it show up in CSS computed values as a matrix">
		<style>
			#box{
				margin-top:30px;
				display: block;
				width: 50px;
				height: 50px;
				background-color: green;
			}
		</style>
	</head>
    <body>
		<h1>Rotate via JS</h1>
        <div id="log"></div>
        <div id="box"></div>
		<script>
			test(function() {
				var box = document.getElementById("box");
				box.style.transform = "rotate(30deg)";
				assert_equals(box.style.transform, "rotate(30deg)");
				assert_equals(window.getComputedStyle(box).getPropertyValue("transform"),
				              "matrix(0.866025, 0.5, -0.5, 0.866025, 0, 0)");
			});
		</script>
    </body>
</html>
